[OT148-85] modified CategoryController
